更新 Page#screencast方法，帧率更加平衡(2024.10.16)

+++ src/main/java/com/ruiyun/jvppeteer/entities/ScreencastFrameMetadata.java
@@ -1,5 +1,7 @@
 package com.ruiyun.jvppeteer.entities;
 
+import java.math.BigDecimal;
+
 public class ScreencastFrameMetadata {
 
     /**
@@ -29,7 +31,7 @@ public class ScreencastFrameMetadata {
     /**
      * Frame swap timestamp.
      */
-    private Long timestamp;
+    private BigDecimal timestamp;
 
     public double getOffsetTop() {
         return offsetTop;
@@ -79,11 +81,11 @@ public class ScreencastFrameMetadata {
         this.scrollOffsetY = scrollOffsetY;
     }
 
-    public Long getTimestamp() {
+    public BigDecimal getTimestamp() {
         return timestamp;
     }
 
-    public void setTimestamp(Long timestamp) {
+    public void setTimestamp(BigDecimal timestamp) {
         this.timestamp = timestamp;
     }
